Replace unsigned and int with size_t in appropriate places.

+++ src/shuffle.cc
@@ -16,12 +16,12 @@ void ShuffleChain::Clear() {
 
 void ShuffleChain::Add(ShuffleItem item) {
     _items.emplace_back(item);
-    _pool.push_back(static_cast<int>(_items.size()) - 1);
+    _pool.push_back(_items.size() - 1);
 }
 
-unsigned ShuffleChain::Len() { return _items.size(); }
-unsigned ShuffleChain::LenURIs() {
-    unsigned sum = 0;
+size_t ShuffleChain::Len() { return _items.size(); }
+size_t ShuffleChain::LenURIs() {
+    size_t sum = 0;
     for (auto& group : _items) {
         sum += group._uris.size();
     }
@@ -32,7 +32,7 @@ unsigned ShuffleChain::LenURIs() {
 void ShuffleChain::FillWindow() {
     while (_window.size() <= _max_window && _pool.size() > 0) {
         /* push a random song from the pool onto the end of the window */
-        unsigned idx = rand() % _pool.size();
+        size_t idx = rand() % _pool.size();
         _window.push_back(_pool[idx]);
         _pool.erase(_pool.begin() + idx);
     }
@@ -41,7 +41,7 @@ void ShuffleChain::FillWindow() {
 const std::vector<std::string>& ShuffleChain::Pick() {
     assert(Len() != 0 && "cannot pick from empty chain");
     FillWindow();
-    int picked_idx = _window[0];
+    size_t picked_idx = _window[0];
     _window.pop_front();
     _pool.push_back(picked_idx);
     return _items[picked_idx]._uris;
